==> differential_binding_analyses/12_DiffBind.Rmd <==
---
title: "differential peak analysis"
author: "Cynthia Zhang"
date: "2023-06-22"
output: html_document
---

```{r}
# The following example code computes differentially bound sites of a pair of conditions using the DiffBind library
# INPUT: bam files from 1_bam_dedup_unmerged.sh, narrowpeak files from 4_macs2_call_peaks.sh
# OUTPUT: differential peaks tsv file, volcano plot (see example output png file)
```

```{r}
#if (!require("BiocManager", quietly = TRUE))
#    install.packages("BiocManager")
#BiocManager::install("DiffBind",force = TRUE)
#BiocManager::install("magrittr",type="source",force = TRUE)

library(DiffBind)
library(magrittr)
```


```{r}
# Diffbind for MSL2 clampDH-clampDH vs yw males

# read in input configuration
# the configuration file contains file paths to input files (peaks to compare)
df <- read.csv('path/to/12a_DiffBind_input_config.csv')
sample.df <- dba(sampleSheet=df)

# sanity check 
plot(sample.df)
dba.plotPCA(sample.df)

# run diffbind
DE.clamp_mutant_dataset <- sample.df %>% 
  dba.blacklist(blacklist = TRUE,greylist = FALSE) %>%
  dba.count(summits = 250) %>%
  dba.normalize() %>%
  dba.contrast(contrast = c("Condition", "msl2", "wt")) %>%
  dba.analyze(method = DBA_DESEQ2)

dba.save(DE.clamp_mutant_dataset, '/path/to/MSL2_clampDH-clampDH_vs_yw_males.dba')

# volcano plot
png(file='/path/to/MSL2_clampDH-clampDH_vs_yw_males.png',width=6, height=4, units='in', pointsize = 20, res=350)
dba.plotVolcano(DE.clamp_mutant_dataset)
dev.off()

# differential peaks
report <- dba.report(DE.clamp_mutant_dataset)
report.df <- as.data.frame(report)
write.table(report.df, "/path/to/MSL2_clampDH-clampDH_vs_yw_males.tsv", sep = "\t", quote = FALSE, row.names = FALSE)
```
