==> differential_binding_analyses/13_clean_diffbind_tsv.ipynb <==
Example code for one pair of comparison
Get significant differential binding (FDR<=0.05) and separate each file by chrX/autosomes and up/down regulated
INPUT: DiffBind peaks in tsv files outputted from 12_DiffBind.Rmd
OUTPUT: chrX/autosomes and up/down regulated tsv files, and _clean.tsv files with only the first three columns

import pandas as pd

df = pd.read_csv('/path/to/MSL2_clampDH-clampDH_vs_yw_males.tsv', sep='\t')

df_1 = df[(df['seqnames'].str.contains('chrX')) & ~(df['seqnames'].str.contains('chrX_')) & (df['Fold']>0) & (df['FDR']<=0.05)]
df_1.to_csv('/path/to/macs2/tsv_files/MSL2_clampDH-clampDH_vs_yw_males_chrX_up.tsv', sep="\t", index=False)
df_1.iloc[:,0:3].to_csv('./macs2/tsv_files/MSL2_clampDH-clampDH_vs_yw_males_chrX_up_clean.tsv', sep="\t", index=False, header=False)

df_2 = df[(df['seqnames'].str.contains('chrX')) & ~(df['seqnames'].str.contains('chrX_')) & (df['Fold']<0) & (df['FDR']<=0.05)]
df_2.to_csv('/path/to/macs2/tsv_files/MSL2_clampDH-clampDH_vs_yw_males_chrX_down.tsv', sep="\t", index=False)
df_2.iloc[:,0:3].to_csv('/path/to/macs2/tsv_files/MSL2_clampDH-clampDH_vs_yw_males_chrX_down_clean.tsv', sep="\t", index=False, header=False)

df_3 = df[~(df['seqnames'].str.contains('chrX')) & (df['Fold']>0) & (df['FDR']<=0.05)]
df_3.to_csv('/path/to/macs2/tsv_files/MSL2_clampDH-clampDH_vs_yw_males_autosomes_up.tsv', sep="\t", index=False)
df_3.iloc[:,0:3].to_csv('/path/to/macs2/tsv_files/MSL2_clampDH-clampDH_vs_yw_males_autosomes_up_clean.tsv', sep="\t", index=False, header=False)

df_ = df[~(df['seqnames'].str.contains('chrX')) & (df['Fold']<0) & (df['FDR']<=0.05)]
df_.to_csv('./macs2/tsv_files/MSL2_clampDH-clampDH_vs_yw_males_autosomes_down.tsv', sep="\t", index=False)
df_.iloc[:,0:3].to_csv('/path/to/macs2/tsv_files/MSL2_clampDH-clampDH_vs_yw_males_autosomes_down_clean.tsv', sep="\t", index=False, header=False)
