==> differential_binding_analyses/14_ChIPseeker.Rmd <==
---
title: "ChIPseeker"
output: html_document
date: "2023-08-15"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
# The following code computes annotate the differential peaks using the ChIPseeker library
# INPUT: differential peaks from 12_DiffBind.Rmd that has been processed by 13_clean_diffbind_tsv.ipynb
# OUTPUT: annotated peaks tsv files and barplots showing composition of different genomic regions (see example output png file)
```

```{r}
# Install and load required packages
# if (!requireNamespace("BiocManager", quietly = TRUE))
#   install.packages("BiocManager")

# BiocManager::install("ChIPseeker")
# BiocManager::install("TxDb.Dmelanogaster.UCSC.dm6.ensGene") 
# BiocManager::install("org.Dm.eg.db")
# BiocManager::install("GO.db")
# BiocManager::install("HDO.db")

library(ChIPseeker)
library(org.Dm.eg.db)
library(GO.db)
library(HDO.db)
library(TxDb.Dmelanogaster.UCSC.dm6.ensGene)
```

```{r}
in_dir <- "/path/to/diffbind/tsv_files/"
out_dir <- "/path/to/output/"
```


```{r}
# csv file containing file names
csv_file <- file.path(in_dir, "1_files_to_annotate.csv")
con <- file(csv_file, "r")

while (length(line <- readLines(con, n = 1)) > 0) {
  tryCatch({
    fields <- strsplit(line, ",")[[1]]
    name <- gsub("\\..*", "", fields[1])
    base_name <- sub("_[^_]*$", "", name)
    
    cat(paste("Working on sample '", base_name, "'\n", sep = ""))
    
    peak_file <- paste(in_dir, fields, sep = "")
    peaks <- import.bed(peak_file)
    
    # Annotate peaks using the dm6 genome
    txdb <- TxDb.Dmelanogaster.UCSC.dm6.ensGene
    peak_anno <- annotatePeak(peaks, tssRegion=c(-3000, 3000), TxDb=txdb, annoDb="org.Dm.eg.db")
    
    # Save annotated peaks
    outfile_name <- paste(out_dir, base_name, ".tsv", sep = "")
    write.table(peak_anno, file = outfile_name, sep = "\t", row.names = FALSE)
    
    # Plot
    p <- plotAnnoBar(peak_anno, title = base_name)
    plot_name <- paste(out_dir, base_name, ".png", sep = "")
    png(file = plot_name, width = 6, height = 3 , units = "in", pointsize = 8, res = 350)
    print(p)
    dev.off()

  }, error = function(e) {
    # Error handling code to skip to the next iter
    cat("Error occurred:", conditionMessage(e), "\n")
    return()
  })
}
```
